fix: implementación y manejo de fechas en mensajes
