+++ 03_Implementation/Juatsapp/src/main/java/org/itson/utils/ChatItem.java
@@ -12,11 +12,7 @@ public final class ChatItem {
      * Nombre del chat.
      */
     private String nombreChat;
-
-    /**
-     * Fecha con formato.
-     */
-    private String fechaTexto;
+    
     /**
      * Imagen del chat.
      */
@@ -59,24 +55,6 @@ public final class ChatItem {
      */
     public void setNombreChat(final String nombreChat) {
         this.nombreChat = nombreChat;
-    }
-
-    /**
-     * Obtiene la fecha con formato.
-     *
-     * @return la fecha con formato.
-     */
-    public String getFechaTexto() {
-        return fechaTexto;
-    }
-
-    /**
-     * Establece la fecha con formato.
-     *
-     * @param fechaTexto
-     */
-    public void setFechaTexto(final String fechaTexto) {
-        this.fechaTexto = fechaTexto;
     }
 
     /**
